I'm going insane, why MAX_DIMENSIONS_EXCEEDED

+++ PercentCoveredByAmbulences/constants.h
@@ -2,10 +2,10 @@
 #define CONSTANTS_H
 
 const extern int EARTH_RADIUS;
-const extern float LOWEST_LAT;
-const extern float LOWEST_LONG;
-const extern float HIGHEST_LAT;
-const extern float HIGHEST_LONG;
+const extern double LOWEST_LAT;
+const extern double LOWEST_LONG;
+const extern double HIGHEST_LAT;
+const extern double HIGHEST_LONG;
 const extern double LONG_IN_1_MILE;
 const extern double LAT_IN_1_MILE;
 const extern double MILES_BETWEEN_POINTS;
